Mobile chat housekeeping
