+++ packages/common/src/store/ui/search-users-modal/selectors.ts
@@ -1,4 +1,7 @@
+import { ID } from 'models/Identifiers'
+import { Status } from 'models/Status'
 import { CommonState } from 'store/reducers'
 
-export const getUserList = (state: CommonState) =>
-  state.ui.searchUsersModal.userList
+export const getUserList = (
+  state: CommonState
+): { userIds: ID[]; status: Status } => state.ui.searchUsersModal.userList
